soccerapi/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
distlib==0.3.1
distro==1.4.0
Django==3.1.1
django-extensions==3.0.8
django-nose==1.4.7
django-redis==4.12.1
djangorestframework==3.11.1
drf-yasg==1.17.1
filelock==3.0.12
html5lib==1.0.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
ipaddr==2.2.0
isort==5.5.1
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
nose==1.3.7
openapi-codec==1.3.2
packaging==20.4
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dotenv==0.14.0
pytoml==0.1.21
pytz==2020.1
redis==3.5.3
requests==2.24.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
